requirements.txt
Django==2.0
django-rest-swagger==2.2.0
djangorestframework==3.10.3
